Generalized AccountInfo to MyAccount
