+++ src/pages/index.js
@@ -3,11 +3,13 @@ import LogIn from "./LogIn";
 import LogOut from "./LogOut";
 import SignUp from './SignUp';
 import Profile from "./Profile";
+import MyAccount from "./MyAcount";
 
 export {
     FrontPage,
     LogIn,
     LogOut,
     SignUp,
-    Profile
+    Profile,
+    MyAccount
 }
